backend/requirements.txt
psycopg2~=2.8.5
Django~=3.1
graphene_django~=2.13
django-graphql-jwt~=0.3
graphene~=2.1.8